+++ doks-gradle-plugin/src/main/kotlin/com/rickbusarow/doks/internal/markdown/MarkdownNode.kt
@@ -15,24 +15,31 @@
 
 package com.rickbusarow.doks.internal.markdown
 
+import com.rickbusarow.doks.internal.trees.depthFirstTraversal
 import org.intellij.lang.annotations.Language
 import org.intellij.markdown.IElementType
-import org.intellij.markdown.MarkdownTokenTypes.Companion.HTML_BLOCK_CONTENT
+import org.intellij.markdown.MarkdownElementTypes
+import org.intellij.markdown.MarkdownTokenTypes
 import org.intellij.markdown.ast.ASTNode
 import org.intellij.markdown.ast.LeafASTNode
 import org.intellij.markdown.ast.getTextInNode
 import org.intellij.markdown.flavours.MarkdownFlavourDescriptor
-import org.intellij.markdown.flavours.commonmark.CommonMarkFlavourDescriptor
+import org.intellij.markdown.flavours.gfm.GFMFlavourDescriptor
 import org.intellij.markdown.parser.MarkdownParser
-import java.util.LinkedList
+import org.jetbrains.kotlin.com.intellij.openapi.util.Key
+import org.jetbrains.kotlin.com.intellij.openapi.util.UserDataHolder
 
 internal data class MarkdownNode(
   val node: ASTNode,
   private val fullText: String,
-  val parent: MarkdownNode?
-) {
+  val parent: MarkdownNode?,
+) : UserDataHolder {
+  val text: String by lazy { node.getTextInNode(fullText).toString() }
+
   val isLeaf: Boolean get() = node is LeafASTNode
-  val text by lazy { node.getTextInNode(fullText).toString() }
+  val isParagraph: Boolean get() = node.type == MarkdownElementTypes.PARAGRAPH
+  val isHtmlBlock: Boolean get() = node.type == MarkdownElementTypes.HTML_BLOCK
+  val isLeafOrParagraph: Boolean get() = isLeaf || isParagraph
   val children: List<MarkdownNode> by lazy {
     node.children
       .map { child ->
@@ -44,7 +51,17 @@ internal data class MarkdownNode(
       }
   }
 
-  val type: IElementType get() = node.type
+  val elementType: IElementType get() = node.type
+
+  private val _userData = mutableMapOf<Key<*>, Any?>()
+  override fun <T : Any?> getUserData(key: Key<T>): T? {
+    @Suppress("UNCHECKED_CAST")
+    return _userData[key] as? T
+  }
+
+  override fun <T : Any?> putUserData(key: Key<T>, value: T?) {
+    _userData[key] = value
+  }
 
   companion object {
     fun from(
@@ -56,7 +73,7 @@ internal data class MarkdownNode(
 
     fun from(
       @Language("markdown") markdown: String,
-      markdownParser: MarkdownParser = MarkdownParser(CommonMarkFlavourDescriptor())
+      markdownParser: MarkdownParser = MarkdownParser(GFMFlavourDescriptor())
     ): MarkdownNode = MarkdownNode(
       node = markdownParser.buildMarkdownTreeFromString(markdown),
       fullText = markdown,
@@ -65,55 +82,107 @@ internal data class MarkdownNode(
   }
 }
 
-internal fun MarkdownNode.isOpeningTag(): Boolean {
+internal fun MarkdownNode?.isEOL(): Boolean = this != null && elementType == MarkdownTokenTypes.EOL
 
-  if (type != HTML_BLOCK_CONTENT) {
-    return false
-  }
-
-  return text.matches(openReg)
+internal fun MarkdownNode.previousSiblings(): Sequence<MarkdownNode> {
+  return parent?.children.orEmpty().asSequence().takeWhile { it != this }
 }
 
-internal fun MarkdownNode.isClosingTag(): Boolean {
-
-  if (type != HTML_BLOCK_CONTENT) {
-    return false
-  }
-
-  return text.matches(closeReg)
+internal fun MarkdownNode.previousSibling(): MarkdownNode? {
+  return previousSiblings().lastOrNull()
 }
 
-internal fun MarkdownNode.depthFirst(): Sequence<MarkdownNode> {
+internal fun MarkdownNode.nextSiblings(): Sequence<MarkdownNode> {
+  return parent?.children.orEmpty().asSequence().dropWhile { it != this }.drop(1)
+}
 
-  val toVisit = LinkedList(children)
+internal fun MarkdownNode.nextSibling(): MarkdownNode? {
+  return nextSiblings().firstOrNull()
+}
 
-  return generateSequence(toVisit::removeFirstOrNull) { node ->
+internal fun MarkdownNode.parents(): Sequence<MarkdownNode> = generateSequence(parent) { it.parent }
 
-    repeat(node.children.lastIndex + 1) {
-      toVisit.addFirst(node.children[node.children.lastIndex - it])
-    }
-    toVisit.removeFirstOrNull()
+internal fun MarkdownNode.countParentLists() = parents()
+  .count { it.elementType == MarkdownElementTypes.LIST_ITEM }
+
+internal fun MarkdownNode?.isListItem(): Boolean {
+  return this != null && elementType == MarkdownElementTypes.LIST_ITEM
+}
+
+internal fun MarkdownNode.isParagraphInListItem(): Boolean {
+  return isParagraph && parent.isListItem()
+}
+
+internal fun MarkdownNode.isParagraphInBlockQuote(): Boolean {
+  return isParagraph && parent.isBlockQuoteElement()
+}
+
+internal fun MarkdownNode.isFirstParagraphInParent(): Boolean {
+  return isParagraph && previousSiblings().none { it.isParagraph }
+}
+
+internal fun MarkdownNode.isListItemDelimiter(): Boolean {
+  return elementType == MarkdownTokenTypes.LIST_BULLET || elementType == MarkdownTokenTypes.LIST_NUMBER
+}
+
+internal fun MarkdownNode.countParentBlockQuotes(): Int {
+  return parents().count { it.isBlockQuoteElement() }
+}
+
+internal fun MarkdownNode?.isBlockQuoteElement(): Boolean {
+  if (this == null) return false
+  return elementType == MarkdownElementTypes.BLOCK_QUOTE
+}
+
+internal fun MarkdownNode?.isBlockQuoteToken(): Boolean {
+  return when {
+    this == null -> false
+    !isLeaf -> false
+    elementType == MarkdownTokenTypes.BLOCK_QUOTE -> true
+    else -> isBlockQuoteTokenInWhiteSpace()
   }
 }
 
-internal inline fun MarkdownNode.depthFirst(
+/**
+ * When a block quote is nested, any angle brackets other than the last one are
+ * incorrectly parsed as WHITE_SPACE, and added as the last node of the parent block quote.
+ */
+internal fun MarkdownNode?.isBlockQuoteTokenInWhiteSpace(): Boolean {
+  return this != null &&
+    elementType == MarkdownTokenTypes.WHITE_SPACE &&
+    text.matches("^>\\s&&[^\\n]+\$".toRegex())
+}
+
+internal fun MarkdownNode.isWhiteSpace(): Boolean {
+  return elementType == MarkdownTokenTypes.WHITE_SPACE
+}
+
+internal fun MarkdownNode.isWhiteSpaceAfterNewLine(): Boolean {
+  return isWhiteSpace() && previousSibling()?.elementType == MarkdownTokenTypes.EOL
+}
+
+internal fun MarkdownNode?.isEolFollowedByParagraph(): Boolean {
+  return when {
+    this == null -> false
+    !isEOL() -> false
+    else -> nextSiblings().any { it.isParagraph }
+  }
+}
+
+internal fun MarkdownNode.firstChildOfTypeOrNull(vararg types: IElementType): MarkdownNode? {
+  return childrenDepthFirst()
+    .firstOrNull { it.elementType in types }
+}
+
+internal fun MarkdownNode.firstChildOfType(vararg types: IElementType): MarkdownNode {
+  return childrenDepthFirst()
+    .first { it.elementType in types }
+}
+
+internal fun MarkdownNode.childrenDepthFirst(): Sequence<MarkdownNode> {
+  return depthFirstTraversal(MarkdownNode::children)
+}
+
+internal inline fun MarkdownNode.childrenDepthFirst(
   crossinline predicate: (MarkdownNode) -> Boolean
-): Sequence<MarkdownNode> {
-
-  val toVisit = LinkedList(children.filter(predicate))
-
-  return generateSequence(toVisit::removeFirstOrNull) { node ->
-
-    if (predicate(node)) {
-      val filtered = node.children.filter(predicate)
-
-      repeat(filtered.lastIndex + 1) {
-        toVisit.addFirst(filtered[filtered.lastIndex - it])
-      }
-
-      toVisit.removeFirstOrNull()
-    } else {
-      null
-    }
-  }
-}
+): Sequence<MarkdownNode> = depthFirstTraversal { children.filter(predicate) }